perf(F1): move the sensors-identity guard into the hook's co-located suite
Round-2 review of c31911738 found the unit guard for the dnd-kit sensors regression (useSensor memoizes on its options object, so inline literals hand DndContext a fresh sensors array per render) sitting in a describe("useSidebarReorderDnd") block inside ProjectList.render-counts.test.tsx. Anyone editing useSidebarReorderDnd.ts runs the co-located useSidebarReorderDnd.test.tsx, which stayed green against a re-inlined hook; the regression only surfaced from a ProjectList-named file when the full @bb/app suite ran.

The case now lives as a third it() in the existing describe("useSidebarReorderDnd") block of useSidebarReorderDnd.test.tsx, with its comment naming ProjectList.render-counts.test.tsx as the end-to-end layer that catches the same thing. The render-counts file keeps only its end-to-end assertion and drops the renderHook and useSidebarReorderDnd imports it no longer uses.

Verified: with the pre-fix hook (c31911738~1, inline useSensor options) swapped in, useSidebarReorderDnd.test.tsx now fails only on the moved case (expected [...] to be [...] // Object.is equality on dndContextProps.sensors; 1 failed, 4 passed) and passes with the hook restored byte-identically; both test files pass (2 files, 6 tests); turbo typecheck for @bb/app passes; oxlint and oxfmt are clean.
